good

+++ lib/model/MessageItemOperateListener.dart
@@ -1,3 +1,6 @@
+
+import 'package:fixnum/src/int64.dart';
+
 abstract class MessageItemOperateListener {
   void onDelete(int position);
   void onCopy(int position);
@@ -6,5 +9,5 @@ abstract class MessageItemOperateListener {
   void onSendLocalMsg(String msg, bool isMe, [String msgType = "MSG_TEXT"]);
   void onPlayVideo(String url);
   void onPlayImage(String url);
-  void onReply(String val, int replyId);
+  void onReply(String val, Int64 replyId);
 }
